Update all deps/more refactoring
