+++ .gitignore
@@ -344,3 +344,5 @@ custom_components
 
 resources/gh_cli/
 .github_token
+
+!.vscode/*
