+++ .pre-commit-config.yaml
@@ -9,13 +9,13 @@ repos:
       - id: sync-additional-dependencies
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.5.2
+    rev: v0.5.5
     hooks:
       - id: ruff-format
       - id: ruff
 
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.10.1
+    rev: v1.11.0
     hooks:
       - id: mypy
         args:
@@ -25,7 +25,7 @@ repos:
           - types-requests
           - types-Pillow
           - types-tzlocal
-          - wg-utilities==3.11.3
+          - wg-utilities==5.16.0
           - types-paho-mqtt
           - python-dotenv==1.0.1
