+++ requirements.txt
@@ -1,18 +1,10 @@
-botocore==1.29.114 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:aeacb03303d6babb0490247158647346d27900fcf89bfb9713db41f1ce1cc93f \
-    --hash=sha256:dd61f445eb53fe906dbde86405915cce61ffa13824a1bb7b826bfc7869a3e628
-jetson-gpio==2.1.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux" \
-    --hash=sha256:8cd4f6d10ff13a75becfa35e2c26a17be27eba4a2459a5e451c278a205f9c757
-jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:02e2e4cc71b5bcab88332eebf907519190dd9e6e82107fa7f83b1003a6252980 \
-    --hash=sha256:90261b206d6defd58fdd5e85f478bf633a2901798906be2ad389150c5c60edbe
-paho-mqtt==2.1.0 ; python_version >= "3.10" and python_version < "4.0" \
+paho-mqtt==2.1.0 ; python_version >= "3.11" and python_version < "4.0" \
     --hash=sha256:12d6e7511d4137555a3f6ea167ae846af2c7357b10bc6fa4f7c3968fc1723834 \
     --hash=sha256:6db9ba9b34ed5bc6b6e3812718c7e06e2fd7444540df2455d2c51bd58808feee
-pigpio==1.78 ; python_version >= "3.10" and python_version < "4.0" \
+pigpio==1.78 ; python_version >= "3.11" and python_version < "4.0" \
     --hash=sha256:81e46f640c4e6342881fa9bbe290dbcd4fc179619dc6591e57a9d4a084dc49fa \
     --hash=sha256:91efa50e4990649da97408a384782d6ccf58342fc59cdfe21ed7a42911569975
-pillow==10.4.0 ; python_version >= "3.10" and python_version < "4.0" \
+pillow==10.4.0 ; python_version >= "3.11" and python_version < "4.0" \
     --hash=sha256:02a2be69f9c9b8c1e97cf2713e789d4e398c751ecfd9967c18d0ce304efbf885 \
     --hash=sha256:030abdbe43ee02e0de642aee345efa443740aa4d828bfe8e2eb11922ea6a21ea \
     --hash=sha256:06b2f7898047ae93fad74467ec3d28fe84f7831370e3c258afa533f81ef7f3df \
@@ -93,18 +85,9 @@ pillow==10.4.0 ; python_version >= "3.10" and python_version < "4.0" \
     --hash=sha256:f7baece4ce06bade126fb84b8af1c33439a76d8a6fd818970215e0560ca28c27 \
     --hash=sha256:ff25afb18123cea58a591ea0244b92eb1e61a1fd497bf6d6384f09bc3262ec3e \
     --hash=sha256:ff337c552345e95702c5fde3158acb0625111017d0e5f24bf3acdb9cc16b90d1
-python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:0123cacc1627ae19ddf3c27a5de5bd67ee4586fbdd6440d9748f8abb483d3e86 \
-    --hash=sha256:961d03dc3453ebbc59dbdea9e4e11c5651520a876d0f4db161e8674aae935da9
-six==1.16.0 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:1e61c37477a1626458e36f7b1d82aa5c9b094fa4802892072e49de9c60c4c926 \
-    --hash=sha256:8abb2f1d86890a2dfb989f9a77cfcfd3e47c2a354b01111771326f8aa26e0254
-tk==0.1.0 ; python_version >= "3.10" and python_version < "4.0" \
+tk==0.1.0 ; python_version >= "3.11" and python_version < "4.0" \
     --hash=sha256:60bc8923d5d35f67f5c6bd93d4f0c49d2048114ec077768f959aef36d4ed97f8 \
     --hash=sha256:703a69ff0d5ba2bd2f7440582ad10160e4a6561595d33457dc6caa79b9bf4930
-urllib3==1.26.19 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:37a0344459b199fce0e80b0d3569837ec6b6937435c5244e7fd73fa6006830f3 \
-    --hash=sha256:3e3d753a8618b86d7de333b4223005f68720bcd6a7d2bcb9fbd2229ec7c1e429
-wg-utilities[devices-dht22]==3.11.3 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:7ca1266c8d9018f4aa05b2b2907f7687c3069393fcf55dfb89f5da1cba68cff6 \
-    --hash=sha256:cb06ded2f26ca5a4effd4c1fbe57ae41308ff6db9203392cee57883eff377511
+wg-utilities[devices-dht22,mqtt]==5.16.0 ; python_version >= "3.11" and python_version < "4.0" \
+    --hash=sha256:18ddb8f65645d2afde27a6612a31ce32e5527ee8954110d723ce0a15d0115439 \
+    --hash=sha256:ba58f5499dcd24f08d86e6f8dc072c76926a5e613405c2587518e01596b4160b
