+++ smart_mini_crt_interface/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+__version__ = "0.0.1"
